Bundle Jinja plugin with Prettier
Just adding the Jinja plugin as another Nix package will not to be
enough. Wrapping Prettier and explicitly linking the plugin works.
Unfortunately, this approach does not work with mentioning the plugin in
Prettier's configuration file. As a consequence, the whole configuration
should be moved into the bundle to have one single location where
configuration takes place.